feat(lsp): nika/semanticDocument — one projector, every surface
The semantic-oracle keystone (mega-plan T2): the canonical graph
projection stops being CLI property and becomes LANGUAGE property —
in its own L0 member.

- **`nika-graph`** (new member · split from nika-schema at the 15k
  cap, the nika-dap/nika-models precedents — member ceremony
  complete: workspace + layer L0 + crate-spec + layer-registry row +
  public-api baseline + coverage-floor row + semver exclude + lore
  navigatrice/cartographer). `GraphDoc`/`Node`/`Edge`/`project()`
  move from `nika-cli/verbs/graph.rs` verbatim; the CLI keeps the
  renderers (mermaid · dot · ascii) and re-exports the projection.
  The source enums are #[non_exhaustive] upstream, so every match
  keeps its fail-loud wildcard (a future verb panics at projection
  time instead of emitting a silently-wrong document).
- **`nika/semanticDocument`** — the vendor-prefixed custom request
  (rust-analyzer lsp_ext convention), capability-gated via
  `experimental.nika.semanticDocument.graphFormat: 1`. Params
  `{uri}`; result `{graph, spans}`: `graph` is the projection
  VERBATIM (in-payload `graph_format: 1` — the spec 03 contract,
  additive spec-first evolution) and `spans` maps task ids to their
  declaring token ranges (presentation wrapper — a canvas links
  nodes back to source without re-scanning YAML).
- Findings → `graph: null` (the CLI skips PLAN for the same reason);
  the request always answers — it is a READ, the judgment lives in
  diagnostics. Structure only: no env values, no secret material.

Law tests: CLI↔LSP byte-equality (`to_value(project()) == graph`),
in-payload version pin, spans on declaring tokens, null-graph on
cyclic/unparseable. Probed live over JSON-RPC: capability advertised
at initialize · graph_format 1 · nodes wave-ordered · byte-equal
with `nika inspect --format json` on the same document.

One oracle, three protocols (CLI · LSP · MCP) — the gopls trajectory,
with the version INSIDE the payload.
