+++ crates/nika-graph/src/lib.rs
@@ -1,0 +1,233 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (C) 2024-2026 SuperNovae Studio <contact@supernovae.studio>
+
+//! The canonical graph projection — `graph_format: 1` (spec 03
+//! §graph-projection). ONE projector, every consumer: the CLI's
+//! `inspect --format json|mermaid|dot`, the LSP's
+//! `nika/semanticDocument`, and any future surface read THIS document.
+//! The envelope is versioned INSIDE the payload; evolution is additive
+//! and spec-first (a new field lands in 03-dag first, then here).
+//!
+//! The source enums are `#[non_exhaustive]` upstream, so every match
+//! carries a fail-loud wildcard: a future verb / gate form / error
+//! action panics HERE at projection time instead of silently emitting
+//! a wrong document — enum and projector ship together.
+//!
+//! Moved from `nika-cli/src/verbs/graph.rs` (2026-07-13 · the semantic
+//! oracle arc): the projection is language truth, not CLI presentation
+//! — the renderers (mermaid · dot · ascii) stay CLI-side.
+
+use serde::Serialize;
+
+use nika_schema::check::CheckReport;
+use nika_schema::raw::{RawAction, RawWorkflow};
+use nika_schema::types::{OnErrorAction, WhenGate};
+
+/// The versioned projection envelope (spec 03 §graph-projection · `graph_format: 1`).
+#[derive(Debug, Serialize)]
+pub struct GraphDoc {
+    /// Envelope version — additive evolution only.
+    pub graph_format: u32,
+    /// Workflow id (kebab-case).
+    pub workflow: String,
+    /// Topologically sorted nodes (wave order · stable).
+    pub nodes: Vec<Node>,
+    /// `depends_on` edges (the closed kind enum grows with the spec).
+    pub edges: Vec<Edge>,
+}
+
+/// One task node — static facts only.
+#[derive(Debug, Serialize)]
+pub struct Node {
+    /// Task id.
+    pub id: String,
+    /// The verb (`infer` · `exec` · `invoke` · `agent`).
+    pub verb: &'static str,
+    /// `invoke:` tool id, when the verb is invoke.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub tool: Option<String>,
+    /// Resolved `<provider>/<model>` (task override → workflow default).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub model: Option<String>,
+    /// The `when:` gate (`"true"`/`"false"` literal or the CEL island).
+    pub when: Option<String>,
+    /// `for_each` fan-out, when present.
+    pub fan_out: Option<FanOut>,
+    /// Per-task capability attribution (`exec:` · `fs.read:` ·
+    /// `fs.write:` · `net.http:` · `tool:` — deterministic, BTree-ordered
+    /// within each family). The un-aggregated voice of the same effect
+    /// walk `infer_permits` folds into the workflow boundary; empty for
+    /// a task with no pinnable effect (bare infer · dynamic net/fs).
+    pub permits: Vec<String>,
+    /// Static cost interval `[min_path, worst_case]` USD — only for
+    /// priced inference tasks (the cost lane's honesty: no price, no
+    /// interval).
+    pub cost_interval: Option<[f64; 2]>,
+    /// `retry.max_attempts`, when a retry policy is declared (spec 05).
+    /// One voice: clients read the DECLARED policy here instead of
+    /// re-parsing the YAML.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub retry_max_attempts: Option<u32>,
+    /// `timeout:` as milliseconds, when declared (spec 03 · the parsed
+    /// duration — unambiguous where the source string form is not).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub timeout_ms: Option<u64>,
+    /// `on_error:` action — `recover` · `skip` · `fail_workflow` (spec 05).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub on_error: Option<&'static str>,
+    /// Declared `output:` binding names, in source order (spec 04) — what
+    /// this task PRODUCES for `${{ tasks.<id>.<name> }}` readers.
+    #[serde(skip_serializing_if = "Vec::is_empty")]
+    pub outputs: Vec<String>,
+}
+
+/// `for_each` fan-out description.
+#[derive(Debug, Serialize)]
+pub struct FanOut {
+    /// `"list"` (literal · known count) or `"expression"` (unknown).
+    pub kind: &'static str,
+    /// Element count for the literal-list form.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub count: Option<u64>,
+}
+
+/// One dependency edge.
+#[derive(Debug, Serialize)]
+pub struct Edge {
+    /// Upstream task id.
+    pub from: String,
+    /// Downstream task id.
+    pub to: String,
+    /// Closed enum — `depends_on` today (spec 03 §graph-projection).
+    pub kind: &'static str,
+}
+
+/// Project a checked workflow into the canonical graph document.
+///
+/// Requires a clean-conformance report: without a valid DAG order there
+/// is no topological node list to project.
+#[must_use]
+pub fn project(wf: &RawWorkflow, report: &CheckReport) -> GraphDoc {
+    let workflow = wf
+        .workflow
+        .as_ref()
+        .map_or_else(|| "workflow".to_owned(), |w| w.value.clone());
+    let default_model = wf.model.as_ref().map(|m| m.value.clone());
+
+    let mut nodes = Vec::with_capacity(wf.tasks.len());
+    for wave in &report.waves {
+        for &index in wave {
+            let task = &wf.tasks[index].value;
+            let (verb, tool, model) = action_facts(&task.action, default_model.as_deref());
+            let cost_interval = report
+                .cost
+                .tasks
+                .iter()
+                .find(|c| c.task == task.id.value)
+                .and_then(|c| c.min_path_usd.zip(c.usd).map(|(min, max)| [min, max]));
+            nodes.push(Node {
+                id: task.id.value.clone(),
+                verb,
+                tool,
+                model,
+                when: task.when.as_ref().map(|w| match &w.value {
+                    WhenGate::Literal(b) => b.to_string(),
+                    WhenGate::Expr(e) => e.clone(),
+                    // #[non_exhaustive] future gate forms name themselves.
+                    other => format!("{other:?}"),
+                }),
+                permits: nika_schema::check::task_permits(task),
+                fan_out: task.for_each.as_ref().map(|f| match &f.value {
+                    nika_schema::raw::ForEachValue::List(items) => FanOut {
+                        kind: "list",
+                        count: items.as_array().map(|a| a.len() as u64),
+                    },
+                    nika_schema::raw::ForEachValue::Expression(_) => FanOut {
+                        kind: "expression",
+                        count: None,
+                    },
+                    #[allow(
+                        clippy::unreachable,
+                        reason = "non_exhaustive future variant — enum and projector ship together; fail loud beats silently-wrong output"
+                    )]
+                    other => {
+                        unreachable!("unknown for_each form: {other:?}")
+                    }
+                }),
+                cost_interval,
+                retry_max_attempts: task.retry.as_ref().map(|r| r.value.max_attempts),
+                timeout_ms: task
+                    .timeout
+                    .as_ref()
+                    .map(|t| u64::try_from(t.value.as_millis()).unwrap_or(u64::MAX)),
+                on_error: task.on_error.as_ref().map(|o| match &o.value.action {
+                    OnErrorAction::Recover(_) => "recover",
+                    OnErrorAction::Skip => "skip",
+                    OnErrorAction::FailWorkflow => "fail_workflow",
+                    #[allow(
+                        clippy::unreachable,
+                        reason = "non_exhaustive future variant — enum and projector ship together; fail loud beats silently-wrong output"
+                    )]
+                    other => unreachable!("unknown on_error action: {other:?}"),
+                }),
+                outputs: task.output.iter().map(|(k, _)| k.value.clone()).collect(),
+            });
+        }
+    }
+
+    let mut edges = Vec::new();
+    for task in &wf.tasks {
+        for dep in &task.value.depends_on {
+            edges.push(Edge {
+                from: dep.value.clone(),
+                to: task.value.id.value.clone(),
+                kind: "depends_on",
+            });
+        }
+    }
+    // Stable edge order: by (from, to) — independent of authoring order.
+    // Then dedup: `depends_on: [x, x]` must not lie about cardinality.
+    edges.sort_by(|a, b| (&a.from, &a.to).cmp(&(&b.from, &b.to)));
+    edges.dedup_by(|a, b| a.from == b.from && a.to == b.to);
+
+    GraphDoc {
+        graph_format: 1,
+        workflow,
+        nodes,
+        edges,
+    }
+}
+
+/// (verb, tool, resolved model) for one action.
+fn action_facts(
+    action: &RawAction,
+    default_model: Option<&str>,
+) -> (&'static str, Option<String>, Option<String>) {
+    match action {
+        RawAction::Infer(infer) => (
+            "infer",
+            None,
+            infer
+                .model
+                .as_ref()
+                .map(|m| m.value.clone())
+                .or_else(|| default_model.map(str::to_owned)),
+        ),
+        RawAction::Exec(_) => ("exec", None, None),
+        RawAction::Invoke(invoke) => ("invoke", Some(invoke.tool.value.clone()), None),
+        RawAction::Agent(agent) => (
+            "agent",
+            None,
+            agent
+                .model
+                .as_ref()
+                .map(|m| m.value.clone())
+                .or_else(|| default_model.map(str::to_owned)),
+        ),
+        #[allow(
+            clippy::unreachable,
+            reason = "non_exhaustive future verb — it must teach this projector its name before it can be drawn"
+        )]
+        other => unreachable!("unknown verb: {other:?}"),
+    }
+}
