+++ crates/nika-lsp/Cargo.toml
@@ -11,6 +11,7 @@ homepage.workspace     = true
 publish                = false  # Internal L4 interface crate (driven by the `nika lsp` subcommand).
 
 [dependencies]
+nika-graph = { path = "../nika-graph", version = "0.102.0" }  # the canonical projection (nika/semanticDocument serves it verbatim)
 nika-catalog = { path = "../nika-catalog", version = "0.102.0", features = ["builtins-transforms", "providers"] } # the tool/model vocabulary completions derive from (born-stale law: never hand-tabled)
 nika-types   = { path = "../nika-types", version = "0.102.0" }  # ExtractMode::ALL — the mode: lane derives from the L0 vocabulary
 nika-schema = { path = "../nika-schema", version = "0.102.0" }  # parse + the ADR-092 check ladder + RawWorkflow (the analysis source)
